initial weather providers finalized and tested
- Fixed bugs with parsing MET data
- Added cloud cover mode finding for METAR data (to filter out stations that might not record cloud cover data and just report clear)
- Fixed wrong types in Dark Sky and OWM models
- Added CLR identifier for cloud coverage
- Implemented bitmap generation and remaining methods for transition from using the CurrentWeather Singleton (yay one singleton killed! one more to go...)
- Fixed bug where android.location.Location object isn't serialized properly, now I just pass the variables we need
- Enabled celsius by default cause I like it that way, fight me
- Added new util method for converting to fahrenheit
- Converted main PixelWatchFace.java class to using the new Weather object and system
- Made observer util class observe until the work being observed is finished
- Added copious amounts of logging

+++ wearApp/src/main/java/com/corvettecole/pixelwatchface/api/darksky/models/Currently.java
@@ -38,7 +38,7 @@ public class Currently {
   private Double dewPoint;
   @SerializedName("humidity")
   @Expose
-  private Integer humidity;
+  private Double humidity;
   @SerializedName("pressure")
   @Expose
   private Double pressure;
@@ -62,7 +62,7 @@ public class Currently {
   private Double visibility;
   @SerializedName("ozone")
   @Expose
-  private Integer ozone;
+  private Double ozone;
 
   public Integer getTime() {
     return time;
@@ -144,11 +144,11 @@ public class Currently {
     this.dewPoint = dewPoint;
   }
 
-  public Integer getHumidity() {
+  public Double getHumidity() {
     return humidity;
   }
 
-  public void setHumidity(Integer humidity) {
+  public void setHumidity(Double humidity) {
     this.humidity = humidity;
   }
 
@@ -208,11 +208,11 @@ public class Currently {
     this.visibility = visibility;
   }
 
-  public Integer getOzone() {
+  public Double getOzone() {
     return ozone;
   }
 
-  public void setOzone(Integer ozone) {
+  public void setOzone(Double ozone) {
     this.ozone = ozone;
   }
