+++ wearApp/src/main/java/com/corvettecole/pixelwatchface/api/darksky/models/Datum_.java
@@ -17,10 +17,10 @@ public class Datum_ {
   private String icon;
   @SerializedName("precipIntensity")
   @Expose
-  private Integer precipIntensity;
+  private Double precipIntensity;
   @SerializedName("precipProbability")
   @Expose
-  private Integer precipProbability;
+  private Double precipProbability;
   @SerializedName("precipType")
   @Expose
   private String precipType;
@@ -85,19 +85,19 @@ public class Datum_ {
     this.icon = icon;
   }
 
-  public Integer getPrecipIntensity() {
+  public Double getPrecipIntensity() {
     return precipIntensity;
   }
 
-  public void setPrecipIntensity(Integer precipIntensity) {
+  public void setPrecipIntensity(Double precipIntensity) {
     this.precipIntensity = precipIntensity;
   }
 
-  public Integer getPrecipProbability() {
+  public Double getPrecipProbability() {
     return precipProbability;
   }
 
-  public void setPrecipProbability(Integer precipProbability) {
+  public void setPrecipProbability(Double precipProbability) {
     this.precipProbability = precipProbability;
   }
